Separated Quantity Modal Component to a different file
